feat(travel): Add selected travel ID in view model and store it during fragment navigation
This commit introduces the `selectedTravelId` LiveData within the `TravelViewModel` to store the ID of the currently selected travel. It also implements the `setSelectedTravel` function to update this ID.
In `DetailFragment`, the `travelId` from the arguments is now stored in the view model when a new travel is loaded.
Additionally this commit changes the instantiation of the ViewModel.