Refacto(app): Refactor ImageUploadManager to use Application context
The ImageUploadManager has been refactored to use the application context instead of the activity context, ensuring that the manager's lifecycle is tied to the application's lifecycle. The `ImageUploadManager` instance is now managed through the `RunningApp` class, which is the application class.
This change also removes the `getInstance()` method from `ImageUploadManager` and manages it's instanciation from the `RunningApp` class.
The upload notification manager now uses the application context as well.
This allows to remove the context when a broadcast is send.
The `UploadImageWorker` now uses the instance of the `ImageUploadManager` provided by `RunningApp`.
The `UploadFragment` now gets the instance of the `ImageUploadManager` from the `RunningApp`.